Make a new module called util and move some functions into it
For tidiness.
