+++ src/engine/strat_engine/backstore/blockdevmgr.rs
@@ -15,9 +15,7 @@ use chrono::{DateTime, Duration, Utc};
 use rand::{seq, thread_rng};
 use uuid::Uuid;
 
-use devicemapper::{
-    Bytes, Device, LinearDevTargetParams, LinearTargetParams, Sectors, TargetLine, IEC,
-};
+use devicemapper::{Bytes, Device, Sectors, IEC};
 
 use crate::{
     engine::{
@@ -97,73 +95,6 @@ impl Recordable<Vec<BaseDevSave>> for Vec<BlkDevSegment> {
             })
             .collect::<Vec<_>>()
     }
-}
-
-/// Append the second list of BlkDevSegments to the first, or if the last
-/// segment of the first argument is adjacent to the first segment of the
-/// second argument, merge those two together.
-/// Postcondition: left.len() + right.len() - 1 <= result.len()
-/// Postcondition: result.len() <= left.len() + right.len()
-// FIXME: There is a method that duplicates this algorithm called
-// coalesce_segs. These methods should either be unified into a single method
-// OR one should go away entirely in solution to:
-// https://github.com/stratis-storage/stratisd/issues/762.
-pub fn coalesce_blkdevsegs(left: &[BlkDevSegment], right: &[BlkDevSegment]) -> Vec<BlkDevSegment> {
-    if left.is_empty() {
-        return right.to_vec();
-    }
-    if right.is_empty() {
-        return left.to_vec();
-    }
-
-    let mut segments = Vec::with_capacity(left.len() + right.len());
-    segments.extend_from_slice(left);
-
-    // Last existing and first new may be contiguous.
-    let coalesced = {
-        let right_first = right.first().expect("!right.is_empty()");
-        let left_last = segments.last_mut().expect("!left.is_empty()");
-        if left_last.uuid == right_first.uuid
-            && (left_last.segment.start + left_last.segment.length == right_first.segment.start)
-        {
-            left_last.segment.length += right_first.segment.length;
-            true
-        } else {
-            false
-        }
-    };
-
-    if coalesced {
-        segments.extend_from_slice(&right[1..]);
-    } else {
-        segments.extend_from_slice(right);
-    }
-    segments
-}
-
-/// Build a linear dev target table from BlkDevSegments. This is useful for
-/// calls to the devicemapper library.
-pub fn map_to_dm(bsegs: &[BlkDevSegment]) -> Vec<TargetLine<LinearDevTargetParams>> {
-    let mut table = Vec::new();
-    let mut logical_start_offset = Sectors(0);
-
-    let segments = bsegs
-        .iter()
-        .map(|bseg| bseg.to_segment())
-        .collect::<Vec<_>>();
-    for segment in segments {
-        let (physical_start_offset, length) = (segment.start, segment.length);
-        let params = LinearTargetParams::new(segment.device, physical_start_offset);
-        let line = TargetLine::new(
-            logical_start_offset,
-            length,
-            LinearDevTargetParams::Linear(params),
-        );
-        table.push(line);
-        logical_start_offset += length;
-    }
-
-    table
 }
 
 #[derive(Debug)]
